refactor: flatten the package to shared modules plus a model tree
Removes the core/, models/, and publish/ wrappers. The shared protocol
machinery sits at the package root and only models are nested, task then
model, so the tree matches deposits/kmesh/qrf95/ exactly.

Refs #1
