+++ docs/training/kmesh/qrf95.md
@@ -1,7 +1,7 @@
---8<-- "src/goldilocks_ml/models/kmesh/qrf95/README.md"
+--8<-- "src/goldilocks_ml/kmesh/qrf95/README.md"
 
 ## Protocol
 
 ```toml
---8<-- "src/goldilocks_ml/models/kmesh/qrf95/protocol.toml"
+--8<-- "src/goldilocks_ml/kmesh/qrf95/protocol.toml"
 ```
