+++ docs/training/metallicity/cgcnn.md
@@ -1,7 +1,7 @@
---8<-- "src/goldilocks_ml/models/metallicity/cgcnn/README.md"
+--8<-- "src/goldilocks_ml/metallicity/cgcnn/README.md"
 
 ## Protocol
 
 ```toml
---8<-- "src/goldilocks_ml/models/metallicity/cgcnn/protocol.toml"
+--8<-- "src/goldilocks_ml/metallicity/cgcnn/protocol.toml"
 ```
