--- src/goldilocks_ml/core/__init__.py
@@ -1,3 +1,0 @@
-"""Reusable training protocol machinery, independent of any particular model."""
-
-from __future__ import annotations
