+++ src/goldilocks_ml/metallicity/cgcnn/README.md
