+++ src/goldilocks_ml/metallicity/cgcnn/__init__.py
