+++ src/goldilocks_ml/metallicity/cgcnn/protocol.toml
