--- src/goldilocks_ml/publish/__init__.py
@@ -1,3 +1,0 @@
-"""Publication of released models to PSDI Data Collections."""
-
-from __future__ import annotations
